First version working.
A first working version of the guessing game from Chapter 2 of the Rust
Book. Attempts to use as little of the following chapters as possible
while still being decent code. The plan is to refactor & add new
features as I learn more about Rust in the following chapters. First up
is refactoring to Modules & Structs, then trying to refactor methods to
Traits as I pursue some sort of inheritance model & extensibility.